Harvest control rule to fish at some fraction of maximum sustainable yield

A simple control rule that specifies the total allowable catch (TAC) as a function of the abundance of the first projection year and some fraction of FMSY/UMSY.

Usage

HCR_MSY(Assessment, reps = 1, MSY_frac = 1, ...)

Arguments

Assessment

An object of class Assessment with estimates of FMSY or UMSY and vulnerable biomass in terminal year.

reps

The number of stochastic samples of the TAC recommendation.

MSY_frac

The fraction of FMSY or UMSY for calculating the TAC (e.g. MSY_frac = 0.75 fishes at 75% of FMSY).

...

Miscellaneous arguments.

Value

An object of class MSEtool::Rec with the TAC recommendation.

Details

The catch advice is calculated using the catch equation of the corresponding assessment. See Assessment@forecast$catch_eq, a function that returns the catch advice for a specified Ftarget.

References

Punt, A. E, Dorn, M. W., and Haltuch, M. A. 2008. Evaluation of threshold management strategies for groundfish off the U.S. West Coast. Fisheries Research 94:251-266.

See also

Author

Q. Huynh

Examples

# create an MP to run in closed-loop MSE (fishes at UMSY)
SPMSY <- make_MP(SP, HCR_MSY)

# The MP which fishes at 75% of FMSY
SP75MSY <- make_MP(SP, HCR_MSY, MSY_frac = 0.75)

# \donttest{
myOM <- MSEtool::runMSE(MSEtool::testOM, MPs = c("FMSYref", "SPMSY", "SP75MSY"))
#> ℹ Checking OM for completeness
#> ✔ Loading operating model
#> ✔ Optimizing for user-specified movement
#> ✔ Calculating MSY reference points for each year
#> ✔ Optimizing for user-specified depletion in last historical year
#> ✔ Calculating historical stock and fishing dynamics
#> ✔ Calculating per-recruit reference points
#> ✔ Calculating B-low reference points
#> ✔ Calculating reference yield - best fixed F strategy
#> ✔ Simulating observed data
#> ✔ Running forward projections
#> ✔ 1 / 3  Running MSE for FMSYref 
#>   |==                                                | 2 % ~00s            |===                                               | 4 % ~00s            |====                                              | 6 % ~00s            |=====                                             | 8 % ~00s            |======                                            | 10% ~01s            |=======                                           | 12% ~00s            |========                                          | 14% ~00s            |=========                                         | 16% ~00s            |==========                                        | 18% ~00s            |===========                                       | 20% ~00s            |============                                      | 22% ~00s            |=============                                     | 24% ~00s            |==============                                    | 27% ~00s            |===============                                   | 29% ~00s            |================                                  | 31% ~00s            |=================                                 | 33% ~00s            |==================                                | 35% ~00s            |===================                               | 37% ~00s            |====================                              | 39% ~00s            |=====================                             | 41% ~00s            |======================                            | 43% ~00s            |=======================                           | 45% ~00s            |========================                          | 47% ~00s            |=========================                         | 49% ~00s            |==========================                        | 51% ~00s            |===========================                       | 53% ~00s            |============================                      | 55% ~00s            |=============================                     | 57% ~00s            |==============================                    | 59% ~00s            |===============================                   | 61% ~00s            |================================                  | 63% ~00s            |=================================                 | 65% ~00s            |==================================                | 67% ~00s            |===================================               | 69% ~00s            |====================================              | 71% ~00s            |=====================================             | 73% ~00s            |======================================            | 76% ~00s            |=======================================           | 78% ~00s            |========================================          | 80% ~00s            |=========================================         | 82% ~00s            |==========================================        | 84% ~00s            |===========================================       | 86% ~00s            |============================================      | 88% ~00s            |=============================================     | 90% ~00s            |==============================================    | 92% ~00s            |===============================================   | 94% ~00s            |================================================  | 96% ~00s            |================================================= | 98% ~00s            |==================================================| 100% elapsed=01s  
#> 
#> ✔ 2 / 3  Running MSE for SPMSY 
#>   |==                                                | 2 % ~04s            |===                                               | 4 % ~02s            |====                                              | 6 % ~02s            |=====                                             | 8 % ~01s            |======                                            | 10% ~02s            |=======                                           | 12% ~02s            |========                                          | 14% ~01s            |=========                                         | 16% ~01s            |==========                                        | 18% ~02s            |===========                                       | 20% ~02s            |============                                      | 22% ~01s            |=============                                     | 24% ~01s            |==============                                    | 27% ~02s            |===============                                   | 29% ~01s            |================                                  | 31% ~01s            |=================                                 | 33% ~01s            |==================                                | 35% ~01s            |===================                               | 37% ~01s            |====================                              | 39% ~01s            |=====================                             | 41% ~01s            |======================                            | 43% ~01s            |=======================                           | 45% ~01s            |========================                          | 47% ~01s            |=========================                         | 49% ~01s            |==========================                        | 51% ~01s            |===========================                       | 53% ~01s            |============================                      | 55% ~01s            |=============================                     | 57% ~01s            |==============================                    | 59% ~01s            |===============================                   | 61% ~01s            |================================                  | 63% ~01s            |=================================                 | 65% ~01s            |==================================                | 67% ~01s            |===================================               | 69% ~01s            |====================================              | 71% ~01s            |=====================================             | 73% ~01s            |======================================            | 76% ~01s            |=======================================           | 78% ~00s            |========================================          | 80% ~00s            |=========================================         | 82% ~00s            |==========================================        | 84% ~00s            |===========================================       | 86% ~00s            |============================================      | 88% ~00s            |=============================================     | 90% ~00s            |==============================================    | 92% ~00s            |===============================================   | 94% ~00s            |================================================  | 96% ~00s            |================================================= | 98% ~00s            |==================================================| 100% elapsed=02s  
#> 
#> ✔ 3 / 3  Running MSE for SP75MSY 
#>   |==                                                | 2 % ~03s            |===                                               | 4 % ~02s            |====                                              | 6 % ~01s            |=====                                             | 8 % ~01s            |======                                            | 10% ~01s            |=======                                           | 12% ~01s            |========                                          | 14% ~01s            |=========                                         | 16% ~01s            |==========                                        | 18% ~01s            |===========                                       | 20% ~01s            |============                                      | 22% ~01s            |=============                                     | 24% ~01s            |==============                                    | 27% ~01s            |===============                                   | 29% ~01s            |================                                  | 31% ~01s            |=================                                 | 33% ~01s            |==================                                | 35% ~01s            |===================                               | 37% ~01s            |====================                              | 39% ~01s            |=====================                             | 41% ~01s            |======================                            | 43% ~01s            |=======================                           | 45% ~01s            |========================                          | 47% ~01s            |=========================                         | 49% ~01s            |==========================                        | 51% ~01s            |===========================                       | 53% ~01s            |============================                      | 55% ~01s            |=============================                     | 57% ~01s            |==============================                    | 59% ~01s            |===============================                   | 61% ~01s            |================================                  | 63% ~01s            |=================================                 | 65% ~01s            |==================================                | 67% ~01s            |===================================               | 69% ~00s            |====================================              | 71% ~00s            |=====================================             | 73% ~00s            |======================================            | 76% ~00s            |=======================================           | 78% ~00s            |========================================          | 80% ~00s            |=========================================         | 82% ~00s            |==========================================        | 84% ~00s            |===========================================       | 86% ~00s            |============================================      | 88% ~00s            |=============================================     | 90% ~00s            |==============================================    | 92% ~00s            |===============================================   | 94% ~00s            |================================================  | 96% ~00s            |================================================= | 98% ~00s            |==================================================| 100% elapsed=02s  
#> 
# }