Segmented harvest control rules

A linear segmented output control rule where the target F (used for the TAC recommendation) is a function of an operational control point (OCP) such as spawning depletion or spawning biomass relative to that at MSY. The joints of the HCR are specified by arguments OCP and relF. Beyond the range of OCP, the response will be flat. HCR_ramp uses HCR_segment with two control points.

HCR_segment(
  Assessment,
  reps = 1,
  OCP_type = c("SSB_SSB0", "SSB_SSBMSY", "SSB_dSSB0", "F_FMSY", "F_F01", "F_FSPR"),
  Ftarget_type = c("FMSY", "F01", "Fmax", "FSPR"),
  OCP = c(0.1, 0.4),
  relF = c(0, 1),
  SPR_OCP,
  SPR_targ,
  ...
)

Arguments

Assessment
An object of class Assessment with estimates of FMSY or UMSY, vulnerable biomass, and spawning biomass depletion in terminal year.

reps
The number of stochastic samples of the TAC recommendation.

OCP_type
The type of operational control points (OCPs) for the harvest control rule used to determine the reduction in F. By default, use ("SSB_SSB0" for spawning depletion. Other biomass OCPs include "SSB_SSBMSY" for spawning biomass relative to MSY and "SSB_dSSB0", for dynamic depletion (dynamic SSB0 is the historical reconstructed biomass with F = 0). For F-based OCPs, the terminal year fishing mortality relative F01 or Fmax (using yield-per-recruit) or F-SPR% (see SPR_OCP argument) can be used.

Ftarget_type
The type of F used for the target fishing mortality rate.

OCP
Numeric vector of operational control points for the HCR (in increasing order).

relF
Numeric vector of Ftarget corresponding to the values in OCP.

SPR_OCP
The value of spawning potential ratio for the OCP if OCP_type = "F_FSPR". By default, 0.4 (F40%).

SPR_targ
The target value of spawning potential ratio if Ftarget_type = "FSPR". By default, 0.4 (F40%).

...
Miscellaneous arguments.

Value

An object of class Rec with the TAC recommendation.

Author

Q. Huynh

Examples

# This is an MP with a 40-10 harvest control rule
DD_40_10 <- make_MP(DD_TMB, HCR_segment, OCP_type = "SSB_SSB0", OCP = c(0.1, 0.4), relF = c(0, 1))