Class-RCModel

An S4 class for the output from RCM.

Slots

OM

An updated operating model, class OM.

SSB

A matrix of estimated spawning biomass with OM@nsim rows and OM@nyears+1 columns.

NAA

An array for the predicted numbers at age with dimension OM@nsim, OM@nyears+1, and OM@maxage+1.

CAA

An array for the predicted catch at age with dimension OM@nsim, OM@nyears, OM@maxage, and nfleet.

CAL

An array for the predicted catch at length with dimension OM@nsim, OM@nyears, length bins, and nfleet.

conv

A logical vector of length OM@nsim indicating convergence of the RCM in the i-th simulation.

Misc

A list of length OM@nsim with more output from the fitted RCM. Within each simulation, items of interest include:

B - total biomass - vector of length nyears+1

E0 - annual unfished spawning biomass - vector of length nyears

E0_SR - unfished spawning biomass for the stock-recruit relationship - numeric

CR - annual compensation ratio - vector of length nyears

Arec - alpha parameter of the stock-recruit relationship - numeric

Brec - beta parameter of the stock-recruit relationship - numeric

R - recruitment - vector of length nyears+1

R_early - recruitment for the cohorts in first year of the model - vector maxage-1

VB - vulnerable biomass - matrix of nyears x nfleet

N - abundance at age - matrix of nyears+1 x maxage

F - apical fishing mortality - matrix of nyears x nfleet

F_at_age - fishing mortality at age - array of nyears x maxage x nfleet

F_equilibrium - equilibrium fishing mortality prior to first year - vector of length nfleet

M - natural mortality - matrix of nyears x maxage

Z - total mortality - matrix of nyears x maxage

q - survey catchability - vector of length nsurvey

s_vul - survey selectivity at age - array of dim nyears+1, maxage, nsurvey

s_vul_len - corresponding survey selectivity at length - matrix of nbins x nsurvey

Ipred - predicted index values - matrix of nyears x nsurvey

s_CAApred - predicted survey catch at age - array of dim nyears, maxage, nsurvey

vul - fleet selectivity at age - array of dim nyears+1, maxage, nfleet (or nsel_block)

vul_len - corresponding fleet selectivity at length - matrix of nbins x nfleet (or nsel_block)

s_CALpred - predicted survey catch at length - array of dim nyears, nbins, nsurvey

MLpred - predicted mean length - matrix of nyears x nfleet

MWpred - predicted mean weight - matrix of nyears x nfleet

CAApred - predicted catch at age - array of nyears, maxage, nfleet

CALpred - predicted catch at length - array of nyears, nbins, nfleet

Cpred - predicted catch in weight - matrix of nyears x nfleet

CN - predicted catch in numbers - matrix of nyears x nfleet

nll - Total objective function of the model - numeric

mean_fit

A list of output from fit to mean values of life history parameters in the operating model. The named list consists of:

obj - a list with components returned from MakeADFun.

opt - a list with components from calling nlminb to obj.

SD - a list (class sdreport) with parameter estimates and their standard errors, obtained from sdreport.

report - a list of model output reported from the TMB executable, i.e. obj$report(). See Misc.

data

A list of the data inputs for the RCM.

config

A data frame describing configuration of the RCM (not currently used).

See also

Author

Q. Huynh