SCA with random walk in M

SCA_RWM is a modification of SCA that incorporates a random walk in M in logit space (constant with age). Set the variance to a small value (0.001) in order to fix M for all years, which is functionally equivalent to SCA.

SCA_RWM(
  x = 1,
  Data,
  AddInd = "B",
  SR = c("BH", "Ricker"),
  vulnerability = c("logistic", "dome"),
  CAA_dist = c("multinomial", "lognormal"),
  CAA_multiplier = 50,
  rescale = "mean1",
  max_age = Data@MaxAge,
  start = NULL,
  fix_h = TRUE,
  fix_F_equilibrium = TRUE,
  fix_omega = TRUE,
  fix_tau = TRUE,
  LWT = NULL,
  early_dev = c("comp_onegen", "comp", "all"),
  late_dev = "comp50",
  refyear = expression(length(Data@Year)),
  M_bounds = c(0.5, 2),
  integrate = FALSE,
  silent = TRUE,
  opt_hess = FALSE,
  n_restart = ifelse(opt_hess, 0, 1),
  control = list(iter.max = 2e+05, eval.max = 4e+05),
  inner.control = list(),
  ...
)

Arguments

x
A position in the Data object (by default, equal to one for assessments).

Data
An object of class Data

AddInd
A vector of integers or character strings indicating the indices to be used in the model. Integers assign the index to the corresponding index in Data@AddInd, "B" (or 0) represents total biomass in Data@Ind, "VB" represents vulnerable biomass in Data@VInd, and "SSB" represents spawning stock biomass in Data@SpInd.

SR
Stock-recruit function (either "BH" for Beverton-Holt or "Ricker").

vulnerability
Whether estimated vulnerability is "logistic" or "dome" (double-normal). See details for parameterization.

CAA_dist
Whether a multinomial or lognormal distribution is used for likelihood of the catch-at-age matrix. See details.

CAA_multiplier
Numeric for data weighting of catch-at-age matrix if CAA_hist = "multinomial". Otherwise ignored. See details.

rescale
A multiplicative factor that rescales the catch in the assessment model, which can improve convergence. By default, "mean1" scales the catch so that time series mean is 1, otherwise a numeric. Output is re-converted back to original units.

max_age
Integer, the maximum age (plus-group) in the model.

start
Optional list of starting values. Entries can be expressions that are evaluated in the function. See details.

fix_h
Logical, whether to fix steepness to value in Data@steep in the model for SCA. This only affects calculation of reference points for SCA2.

fix_F_equilibrium
Logical, whether the equilibrium fishing mortality prior to the first year of the model is estimated. If TRUE, F_equilibrium is fixed to value provided in start (if provided), otherwise, equal to zero (assumes unfished conditions).

fix_omega
Logical, whether the standard deviation of the catch is fixed. If TRUE, sigma is fixed to value provided in start (if provided), otherwise, value based on Data@CV_Cat.

fix_tau
Logical, the standard deviation of the recruitment deviations is fixed. If TRUE, tau is fixed to value provided in start (if provided), otherwise, value based on Data@sigmaR.

LWT
A vector of likelihood weights for each survey.

early_dev
Numeric or character string describing the years for which recruitment deviations are estimated in SCA. By default, equal to "comp_onegen", where rec devs are estimated one full generation prior to the first year when catch-at-age (CAA) data are available. With "comp", rec devs are estimated starting in the first year with CAA. With "all", rec devs start at the beginning of the model. If numeric, the number of years after the first year of the model for which to start estimating rec devs. Use negative numbers for years prior to the first year.

late_dev
Typically, a numeric for the number of most recent years in which recruitment deviations will not be estimated in SCA (recruitment in these years will be based on the mean predicted by stock-recruit relationship). By default, "comp50" uses the number of ages (smaller than the mode) for which the catch-at-age matrix has less than half the abundance than that at the mode.

refyear
An expression for the year for which M is used to report MSY and depletion reference points. By default, terminal year.

M_bounds
A numeric vector of length 2 to indicate the minimum and maximum M in the random walk as a proportion of the starting M (M_start). The default min and max are 50% and 200%, respectively.

integrate
Logical, whether the likelihood of the model integrates over the likelihood of the recruitment deviations and M random walk (thus, treating it as a random effects/state-space variable). Otherwise, recruitment deviations and the random walk are penalized parameters.

silent
Logical, passed to MakeADFun, whether TMB will print trace information during optimization. Used for diagnostics for model convergence.

opt_hess
Logical, whether the hessian function will be passed to nlminb during optimization (this generally reduces the number of iterations to convergence, but is memory and time intensive and does not guarantee an increase in convergence rate). Ignored if integrate = TRUE.

n_restart
The number of restarts (calls to nlminb) in the optimization procedure, so long as the model hasn't converged. The optimization continues from the parameters from the previous (re)start.

control
A named list of arguments for optimization to be passed to nlminb.

inner.control
A named list of arguments for optimization of the random effects, which is passed on to newton.

...
Other arguments to be passed.

Value

An object of class Assessment.

Details

The model estimates year-specific M (constant with age) as a random walk in logit space, bounded by a proportion of M_start (specified in M_bounds).

The starting value for the first year M (M_start) is Data@Mort[x] and is fixed. The fixed SD of the random walk (tau_M) is 0.05.

Alternative values can be provided in the start list (see example).

See SCA for all other information about the structure and setup of the model.

The SCA builds in a stock-recruit relationship into the model. Annual unfished and MSY reference points are calculated and reported in TMB_report.

See also

Author

Q. Huynh

Examples

res <- SCA_RWM(Data = MSEtool::SimulatedData, start = list(M_start = 0.4, tau_M = 0.05))

#> Error in if (obj$env$data$catch_eq == "Pope") {    dots <- list(...)    dots$data$C_hist <- rep(1e-08, dots$data$n_y)    par[names(par) == "U_equilibrium"] <- 0    obj2 <- MakeADFun(data = dots$data, parameters = dots$params,         map = dots$map, random = obj$env$random, DLL = "SAMtool",         silent = TRUE)    out <- obj2$report(par)$E} else {    par[names(par) == "log_F_dev"] <- log(1e-08)    par[names(par) == "F_equilibrium"] <- 0    out <- obj$report(par)$E}: argument is of length zero

res2 <- SCA(Data = MSEtool::SimulatedData)
res3 <- SCA_RWM(Data = MSEtool::SimulatedData, start = list(M_start = 0.4, tau_M = 0.001))

#> Error in if (obj$env$data$catch_eq == "Pope") {    dots <- list(...)    dots$data$C_hist <- rep(1e-08, dots$data$n_y)    par[names(par) == "U_equilibrium"] <- 0    obj2 <- MakeADFun(data = dots$data, parameters = dots$params,         map = dots$map, random = obj$env$random, DLL = "SAMtool",         silent = TRUE)    out <- obj2$report(par)$E} else {    par[names(par) == "log_F_dev"] <- log(1e-08)    par[names(par) == "F_equilibrium"] <- 0    out <- obj$report(par)$E}: argument is of length zero

 
# \donttest{
compare_models(res, res2, res3)

#> Error in compare_models(res, res2, res3): object 'res' not found

# }