Model-based management procedures

A suite of model-based management procedures (MPs) included in the package. Additional MPs, with specific model configurations (e.g., stock-recruit function or fixing certain parameters) or alternative ramped harvest control rules can be created with make_MP and the available Assess and HCR objects with constant TAC between assessment years.

SCA_MSY(x, Data, reps = 1, diagnostic = "min")

SCA_75MSY(x, Data, reps = 1, diagnostic = "min")

SCA_4010(x, Data, reps = 1, diagnostic = "min")

DDSS_MSY(x, Data, reps = 1, diagnostic = "min")

DDSS_75MSY(x, Data, reps = 1, diagnostic = "min")

DDSS_4010(x, Data, reps = 1, diagnostic = "min")

SP_MSY(x, Data, reps = 1, diagnostic = "min")

SP_75MSY(x, Data, reps = 1, diagnostic = "min")

SP_4010(x, Data, reps = 1, diagnostic = "min")

SSS_MSY(x, Data, reps = 1, diagnostic = "min")

SSS_75MSY(x, Data, reps = 1, diagnostic = "min")

SSS_4010(x, Data, reps = 1, diagnostic = "min")

Arguments

x

A position in the Data object.

Data

An object of class Data

reps

Numeric, the number of stochastic replicates for the management advice.

diagnostic

Character string describing the assessment diagnostic to save, see make_MP.

Value

An object of class Rec which contains the management recommendation.

Functions

SCA_MSY(): A statistical catch-at-age model with a TAC recommendation based on fishing at FMSY, and default arguments for configuring SCA.

SCA_75MSY(): An SCA with a TAC recommendation based on fishing at 75% of FMSY.

SCA_4010(): An SCA with a 40-10 control rule.

DDSS_MSY(): A state-space delay difference model with a TAC recommendation based on fishing at FMSY, and default arguments for configuring DD_SS.

DDSS_75MSY(): A state-space delay difference model with a TAC recommendation based on fishing at 75% of FMSY.

DDSS_4010(): A state-space delay difference model with a 40-10 control rule.

SP_MSY(): A surplus production model with a TAC recommendation based on fishing at FMSY, and default arguments for configuring SP.

SP_75MSY(): A surplus production model with a TAC recommendation based on fishing at 75% of FMSY.

SP_4010(): A surplus production model with a 40-10 control rule.

SSS_MSY(): Simple stock synthesis (terminal depletion fixed to 0.4 in SSS) with a TAC recommendation based on fishing at FMSY.

SSS_75MSY(): Simple stock synthesis (terminal depletion fixed to 0.4) with with a TAC recommendation based on fishing at 75% FMSY.

SSS_4010(): Simple stock synthesis (terminal depletion fixed to 0.4) with a 40-10 control rule.

Examples

MSEtool::avail("MP", package = "SAMtool")
#> ✔ Searching for objects of class  MP  in package:  SAMtool
#>  [1] "DDSS_4010"  "DDSS_75MSY" "DDSS_MSY"   "SCA_4010"   "SCA_75MSY" 
#>  [6] "SCA_MSY"    "SP_4010"    "SP_75MSY"   "SP_MSY"     "SSS_4010"  
#> [11] "SSS_75MSY"  "SSS_MSY"   

# \donttest{
myMSE <- MSEtool::runMSE(MSEtool::testOM, MPs = c("FMSYref", "SCA_4010"))
#> ✔ Checking MPs
#> ℹ Checking OM for completeness
#> ✔ Loading operating model
#> ✔ Optimizing for user-specified movement
#> ✔ Calculating MSY reference points for each year
#> ✔ Optimizing for user-specified depletion in last historical year
#> ✔ Calculating historical stock and fishing dynamics
#> ✔ Calculating per-recruit reference points
#> ✔ Calculating B-low reference points
#> ✔ Calculating reference yield - best fixed F strategy
#> ✔ Simulating observed data
#> ✔ Running forward projections
#> ✔ 1 / 2  Running MSE for FMSYref 
#>   |                                                          |=                                                 |   2%  |                                                          |==                                                |   4%  |                                                          |===                                               |   6%  |                                                          |====                                              |   8%  |                                                          |=====                                             |  10%  |                                                          |======                                            |  12%  |                                                          |=======                                           |  14%  |                                                          |========                                          |  16%  |                                                          |=========                                         |  18%  |                                                          |==========                                        |  20%  |                                                          |===========                                       |  22%  |                                                          |============                                      |  24%  |                                                          |=============                                     |  27%  |                                                          |==============                                    |  29%  |                                                          |===============                                   |  31%  |                                                          |================                                  |  33%  |                                                          |=================                                 |  35%  |                                                          |==================                                |  37%  |                                                          |===================                               |  39%  |                                                          |====================                              |  41%  |                                                          |=====================                             |  43%  |                                                          |======================                            |  45%  |                                                          |=======================                           |  47%  |                                                          |========================                          |  49%  |                                                          |==========================                        |  51%  |                                                          |===========================                       |  53%  |                                                          |============================                      |  55%  |                                                          |=============================                     |  57%  |                                                          |==============================                    |  59%  |                                                          |===============================                   |  61%  |                                                          |================================                  |  63%  |                                                          |=================================                 |  65%  |                                                          |==================================                |  67%  |                                                          |===================================               |  69%  |                                                          |====================================              |  71%  |                                                          |=====================================             |  73%  |                                                          |======================================            |  76%  |                                                          |=======================================           |  78%  |                                                          |========================================          |  80%  |                                                          |=========================================         |  82%  |                                                          |==========================================        |  84%  |                                                          |===========================================       |  86%  |                                                          |============================================      |  88%  |                                                          |=============================================     |  90%  |                                                          |==============================================    |  92%  |                                                          |===============================================   |  94%  |                                                          |================================================  |  96%  |                                                          |================================================= |  98%  |                                                          |==================================================| 100%
#> 
#> ✔ 2 / 2  Running MSE for SCA_4010 
#>   |                                                          |=                                                 |   2%  |                                                          |==                                                |   4%  |                                                          |===                                               |   6%  |                                                          |====                                              |   8%  |                                                          |=====                                             |  10%  |                                                          |======                                            |  12%  |                                                          |=======                                           |  14%  |                                                          |========                                          |  16%  |                                                          |=========                                         |  18%  |                                                          |==========                                        |  20%  |                                                          |===========                                       |  22%  |                                                          |============                                      |  24%  |                                                          |=============                                     |  27%  |                                                          |==============                                    |  29%  |                                                          |===============                                   |  31%  |                                                          |================                                  |  33%  |                                                          |=================                                 |  35%  |                                                          |==================                                |  37%  |                                                          |===================                               |  39%  |                                                          |====================                              |  41%  |                                                          |=====================                             |  43%  |                                                          |======================                            |  45%  |                                                          |=======================                           |  47%  |                                                          |========================                          |  49%  |                                                          |==========================                        |  51%  |                                                          |===========================                       |  53%  |                                                          |============================                      |  55%  |                                                          |=============================                     |  57%  |                                                          |==============================                    |  59%  |                                                          |===============================                   |  61%  |                                                          |================================                  |  63%  |                                                          |=================================                 |  65%  |                                                          |==================================                |  67%  |                                                          |===================================               |  69%  |                                                          |====================================              |  71%  |                                                          |=====================================             |  73%  |                                                          |======================================            |  76%  |                                                          |=======================================           |  78%  |                                                          |========================================          |  80%  |                                                          |=========================================         |  82%  |                                                          |==========================================        |  84%  |                                                          |===========================================       |  86%  |                                                          |============================================      |  88%  |                                                          |=============================================     |  90%  |                                                          |==============================================    |  92%  |                                                          |===============================================   |  94%  |                                                          |================================================  |  96%  |                                                          |================================================= |  98%  |                                                          |==================================================| 100%
#> 
# }