Class-RCModel

An S4 class for the output from RCM.

Slots

OM

An updated operating model, class OM.

SSB

A matrix of estimated spawning biomass with OM@nsim rows and OM@nyears+1 columns.

NAA

An array for the predicted numbers at age with dimension OM@nsim, OM@nyears+1, and OM@maxage+1.

CAA

An array for the predicted catch at age with dimension OM@nsim, OM@nyears, OM@maxage, and nfleet.

CAL

An array for the predicted catch at length with dimension OM@nsim, OM@nyears, length bins, and nfleet.

conv

A logical vector of length OM@nsim indicating convergence of the RCM in the i-th simulation.

Misc

A list of length OM@nsim with more output from the fitted RCM. Within each simulation, a named list containing items of interest include:

B - total biomass - vector of length nyears+1

E0 - annual unfished spawning biomass - vector of length nyears

E0_SR - unfished spawning biomass for the stock-recruit relationship - numeric

CR_SR - compensation ratio correpsponding to Arec and unfished spawning biomass for the stock-recruit relationship - numeric

CR - annual compensation ratio - vector of length nyears

Arec - alpha parameter of the stock-recruit relationship - numeric

Brec - beta parameter of the stock-recruit relationship - numeric

R - recruitment - vector of length nyears+1

R_early - recruitment for the cohorts in first year of the model - vector n_age-1 (where n_age = maxage + 1)

VB - vulnerable biomass - matrix of nyears x nfleet

N - abundance at age - matrix of nyears+1 x n_age

F - apical fishing mortality - matrix of nyears x nfleet

F_at_age - fishing mortality at age - matrix of nyears x n_age

F_equilibrium - equilibrium fishing mortality prior to first year - vector of length nfleet

M - natural mortality - matrix of nyears x n_age

Z - total mortality - matrix of nyears x n_age

q - index catchability - vector of length nsurvey

ivul - index selectivity at age - array of dim nyears+1, n_age, nsurvey

ivul_len - corresponding index selectivity at length - matrix of nbins x nsurvey

Ipred - predicted index values - matrix of nyears x nsurvey

IAApred - predicted index catch at age - array of dim nyears, n_age, nsurvey

vul - fleet selectivity at age - array of dim nyears+1, n_age, nfleet (or nsel_block)

vul_len - corresponding fleet selectivity at length - matrix of nbins x nfleet (or nsel_block)

IALpred - predicted index catch at length - array of dim nyears, nbins, nsurvey

MLpred - predicted mean length - matrix of nyears x nfleet

MWpred - predicted mean weight - matrix of nyears x nfleet

CAApred - predicted catch at age - array of nyears, n_age, nfleet

CALpred - predicted catch at length - array of nyears, nbins, nfleet

Cpred - predicted catch in weight - matrix of nyears x nfleet

CN - predicted catch in numbers - matrix of nyears x nfleet

dynamic_SSB0 - the dynamic unfished spawning biomass calcaluated by projecting the historical model with zero catches - vector of length nyears+1

SPR_eq - equilibrium spawning potential ratio calculated from annual F-at-age - vector of length nyears

SPR_dyn - dynamic (transitional) spawning potential ratio calculated from cumulative survival of cohorts - vector of length nyears

nll - total objective function of the model - numeric

nll_fleet - objective function values for each annual data point(s) from fleets - array of nyears x nfleet x 5 (for Catch, equilibrium catch, CAA, CAL, and mean size)

nll_index - objective function values for each annual data point(s) in the index - array of nyears x nsurvey x 3 (for Index, IAA, and IAL)

prior - penalty value added to the objective function from priors - numeric

penalty - additional penalty values added to the objective function due to high F - numeric

conv - whether the model converged (whether a positive-definite Hessian was obtained) - logical

mean_fit

A list of output from fit to mean values of life history parameters in the operating model. The named list consists of:

obj - a list with components returned from MakeADFun.

opt - a list with components from calling nlminb to obj.

SD - a list (class sdreport) with parameter estimates and their standard errors, obtained from sdreport.

report - a list of model output reported from the TMB executable, i.e. obj$report(). See Misc.

data

A RCMdata object containing data inputs for the RCM.

config

A list describing configuration of the RCM:

drop_sim - a vector of simulations that were dropped for the output

See also

Author

Q. Huynh