Stock Assessment Methods Toolkit

Simulation tools for closed-loop simulation are provided for the 'MSEtool' operating model to inform data-rich fisheries. SAMtool provides an OM conditioning model, assessment models of varying complexity with standardized reporting, diagnostic tools for evaluating assessments within closed-loop simulation, and helper functions for building more complex operating models and model-based management procedures.

How to use SAMtool

The main features of SAMtool are the assessment models and the ability to make model-based management procedures by combining assessment models with harvest control rules. Such MPs can be used and tested in management strategy evaluation with MSEtool operating models. An overview of these features is available in the SAMtool vignette.

The following assessment models are available:

Surplus production (SP, SP_SS, SP_Fox, and spict)

Delay difference (DD_TMB, cDD, DD_SS, and cDD_SS)

Statistical catch-at-age (SCA, SCA2, and SCA_Pope)

Simple Stock Synthesis (SSS which implements SCA_Pope with fixed depletion assumption)

Virtual population analysis (VPA)

The RCM (Rapid Conditioning Model) can be used to condition operating models from real data. Information can be found here.

All SAMtool vignettes can also be viewed by typing browseVignettes("SAMtool") into the R console or through the SAMtool webpage on CRAN.

References

Carruthers, T.R., Punt, A.E., Walters, C.J., MacCall, A., McAllister, M.K., Dick, E.J., Cope, J. 2014. Evaluating methods for setting catch limits in data-limited fisheries. Fisheries Research. 153: 48-68.

Carruthers, T.R., Kell, L.T., Butterworth, D.S., Maunder, M.N., Geromont, H.F., Walters, C., McAllister, M.K., Hillary, R., Levontin, P., Kitakado, T., Davies, C.R. Performance review of simple management procedures. ICES Journal of Marine Science. 73: 464-482.

Author

Quang Huynh quang@bluematterscience.com

Tom Carruthers tom@bluematterscience.com

Adrian Hordyk adrian@bluematterscience.com