Continuous Delay-differential assessment model

A catch and index-based assessment model. Compared to the discrete delay-difference (annual time-step in production and fishing), the delay-differential model (cDD) is based on continuous recruitment and fishing mortality within a time-step. The continuous model works much better for populations with high turnover (e.g. high F or M, continuous reproduction). This model is conditioned on catch and fits to the observed index. In the state-space version (cDD_SS), recruitment deviations from the stock-recruit relationship are estimated.

cDD(
  x = 1,
  Data,
  AddInd = "B",
  SR = c("BH", "Ricker"),
  rescale = "mean1",
  MW = FALSE,
  start = NULL,
  prior = list(),
  fix_h = TRUE,
  dep = 1,
  LWT = list(),
  n_itF = 5L,
  silent = TRUE,
  opt_hess = FALSE,
  n_restart = ifelse(opt_hess, 0, 1),
  control = list(iter.max = 5000, eval.max = 10000),
  ...
)

cDD_SS(
  x = 1,
  Data,
  AddInd = "B",
  SR = c("BH", "Ricker"),
  rescale = "mean1",
  MW = FALSE,
  start = NULL,
  prior = list(),
  fix_h = TRUE,
  fix_sigma = FALSE,
  fix_tau = TRUE,
  dep = 1,
  LWT = list(),
  n_itF = 5L,
  integrate = FALSE,
  silent = TRUE,
  opt_hess = FALSE,
  n_restart = ifelse(opt_hess, 0, 1),
  control = list(iter.max = 5000, eval.max = 10000),
  inner.control = list(),
  ...
)

Arguments

x
An index for the objects in Data when running in closed loop simulation. Otherwise, equals to 1 when running an assessment.

Data
An object of class Data.

AddInd
A vector of integers or character strings indicating the indices to be used in the model. Integers assign the index to the corresponding index in Data@AddInd, "B" (or 0) represents total biomass in Data@Ind, "VB" represents vulnerable biomass in Data@VInd, and "SSB" represents spawning stock biomass in Data@SpInd.

SR
Stock-recruit function (either "BH" for Beverton-Holt or "Ricker").

rescale
A multiplicative factor that rescales the catch in the assessment model, which can improve convergence. By default, "mean1" scales the catch so that time series mean is 1, otherwise a numeric. Output is re-converted back to original units.

MW
Logical, whether to fit to mean weight. In closed-loop simulation, mean weight will be grabbed from Data@Misc[[x]]$MW, otherwise calculated from Data@CAL.

start
Optional list of starting values. Entries can be expressions that are evaluated in the function. See details.

prior
A named list (R0, h, M, and q) to provide the mean and standard deviations of prior distributions for those parameters. R0, index q, and M priors are lognormal (provide the mean in normal space, SD in lognormal space). Beverton-Holt steepness uses a beta prior, while Ricker steepness uses a normal prior. For index q, provide a matrix for nsurvey rows and 2 columns (for mean and SD), with NA in rows corresponding to indices without priors. For all others, provide a length-2 vector for the mean and SD. See vignette for full description.

fix_h
Logical, whether to fix steepness to value in Data@steep in the assessment model.

dep
The initial depletion in the first year of the model. A tight prior is placed on the model objective function to estimate the equilibrium fishing mortality corresponding to the initial depletion. Due to this tight prior, this F should not be considered to be an independent model parameter. Set to zero to eliminate this prior.

LWT
A named list of likelihood weights. For LWT$Index, a vector of likelihood weights for each survey, while for LWT$MW a numeric.

n_itF
Integer, the number of iterations to solve F conditional on the observed catch.

silent
Logical, passed to MakeADFun, whether TMB will print trace information during optimization. Used for diagnostics for model convergence.

opt_hess
Logical, whether the hessian function will be passed to nlminb during optimization (this generally reduces the number of iterations to convergence, but is memory and time intensive and does not guarantee an increase in convergence rate). Ignored if integrate = TRUE.

n_restart
The number of restarts (calls to nlminb) in the optimization procedure, so long as the model hasn't converged. The optimization continues from the parameters from the previous (re)start.

control
A named list of parameters regarding optimization to be passed to nlminb.

...
Additional arguments (not currently used).

fix_sigma
Logical, whether the standard deviation of the index is fixed. If TRUE, sigma is fixed to value provided in start (if provided), otherwise, value based on Data@CV_Ind.

fix_tau
Logical, the standard deviation of the recruitment deviations is fixed. If TRUE, tau is fixed to value provided in start (if provided), otherwise, equal to 1.

integrate
Logical, whether the likelihood of the model integrates over the likelihood of the recruitment deviations (thus, treating it as a state-space variable). Otherwise, recruitment deviations are penalized parameters.

inner.control
A named list of arguments for optimization of the random effects, which is passed on to newton via MakeADFun.

Value

An object of Assessment containing objects and output from TMB.

Details

For start (optional), a named list of starting values of estimates can be provided for:

R0 Unfished recruitment. Otherwise, Data@OM$R0[x] is used in closed-loop, and 400% of mean catch otherwise.

h Steepness. Otherwise, Data@steep[x] is used, or 0.9 if empty.

Kappa Delay-differential Kappa parameter. Otherwise, calculated from biological parameters in the Data object.

F_equilibrium Equilibrium fishing mortality leading into first year of the model (to determine initial depletion). By default, 0.

tau Lognormal SD of the recruitment deviations (process error) for DD_SS. By default, Data@sigmaR[x].

sigma Lognormal SD of the index (observation error). By default, Data@CV_Ind[x]. Not used if multiple indices are used.

sigma_W Lognormal SD of the mean weight (observation error). By default, 0.1.

Multiple indices are supported in the model. Data@Ind, Data@VInd, and Data@SpInd are all assumed to be biomass-based. For Data@AddInd, Data@I_units are used to identify a biomass vs. abundance-based index.

Online Documentation

Model description and equations are available on the openMSE website.

Required Data

cDD: Cat, Ind, Mort, L50, vbK, vbLinf, vbt0, wla, wlb, MaxAge

cDD_SS: Cat, Ind, Mort, L50, vbK, vbLinf, vbt0, wla, wlb, MaxAge

Optional Data

cDD: steep

cDD_SS: steep, CV_Ind, sigmaR

References

Hilborn, R., and Walters, C., 1992. Quantitative Fisheries Stock Assessment: Choice, Dynamics and Uncertainty. Chapman and Hall, New York.

See also

Author

Q. Huynh

Examples

#### Observation-error delay difference model
res <- cDD(Data = MSEtool::Red_snapper)

### State-space version
### Also set recruitment variability SD = 0.6 (since fix_tau = TRUE)
res <- cDD_SS(Data = MSEtool::Red_snapper, start = list(tau = 0.6))

summary(res@SD) # Parameter estimates

#>                 Estimate   Std. Error
#> R0x         -0.480921455 0.0547476197
#> log_sigma   -1.847400042 0.0642471710
#> log_rec_dev  0.081587219 0.6049894072
#> log_rec_dev  0.111204054 0.6125899084
#> log_rec_dev  0.107322308 0.6117275453
#> log_rec_dev  0.097503317 0.6093864955
#> log_rec_dev  0.107647465 0.6124728442
#> log_rec_dev  0.141446939 0.6224339666
#> log_rec_dev  0.180777991 0.6348165576
#> log_rec_dev  0.222165851 0.6488319947
#> log_rec_dev  0.269647741 0.6659351608
#> log_rec_dev  0.317759201 0.6845925330
#> log_rec_dev  0.348397602 0.6979832730
#> log_rec_dev  0.349581550 0.7002855500
#> log_rec_dev  0.339561583 0.6976667607
#> log_rec_dev  0.370494271 0.7111734109
#> log_rec_dev  0.383392138 0.7173261234
#> log_rec_dev  0.401655660 0.7251979535
#> log_rec_dev  0.443006137 0.7430123229
#> log_rec_dev  0.405945004 0.7262746923
#> log_rec_dev  0.354623846 0.7038032205
#> log_rec_dev  0.322122772 0.6897517898
#> log_rec_dev  0.323794867 0.6895276101
#> log_rec_dev  0.328408423 0.6910275805
#> log_rec_dev  0.315127343 0.6863526262
#> log_rec_dev  0.324125255 0.6906325851
#> log_rec_dev  0.369392591 0.7103030309
#> log_rec_dev  0.408297484 0.7296375569
#> log_rec_dev  0.452468887 0.7532735362
#> log_rec_dev  0.526302090 0.7939118642
#> log_rec_dev  0.529252204 0.7976604117
#> log_rec_dev  0.490692480 0.7756016989
#> log_rec_dev  0.511395980 0.7802731004
#> log_rec_dev  0.482631571 0.7577363234
#> log_rec_dev  0.368159321 0.7021732270
#> log_rec_dev  0.241152720 0.6519067220
#> log_rec_dev  0.158679381 0.6230152527
#> log_rec_dev  0.105519680 0.6055344549
#> log_rec_dev  0.081408980 0.5970019161
#> log_rec_dev  0.068901077 0.5920825398
#> log_rec_dev  0.068758365 0.5908982655
#> log_rec_dev  0.036593700 0.5830367106
#> log_rec_dev  0.007575195 0.5762369935
#> log_rec_dev  0.009040124 0.5761802392
#> log_rec_dev  0.045867494 0.5839586492
#> log_rec_dev  0.076353141 0.5905744392
#> log_rec_dev  0.092545589 0.5941710033
#> log_rec_dev  0.074280617 0.5899738527
#> log_rec_dev  0.059466472 0.5864711168
#> log_rec_dev  0.053892245 0.5849228930
#> log_rec_dev  0.053596864 0.5844512081
#> log_rec_dev  0.043762183 0.5817642309
#> log_rec_dev  0.040902993 0.5803827889
#> log_rec_dev  0.051548347 0.5817812078
#> log_rec_dev  0.024813238 0.5752244740
#> log_rec_dev -0.004895483 0.5680683938
#> log_rec_dev -0.007052469 0.5664624794
#> log_rec_dev -0.020608349 0.5623777215
#> log_rec_dev -0.022961957 0.5602817255
#> log_rec_dev -0.054170870 0.5528287600
#> log_rec_dev -0.109250959 0.5416013944
#> log_rec_dev -0.152854059 0.5331731741
#> log_rec_dev -0.162749695 0.5304019677
#> log_rec_dev -0.158932104 0.5300187350
#> log_rec_dev -0.177711307 0.5265857400
#> log_rec_dev -0.223382424 0.5192153469
#> log_rec_dev -0.247060068 0.5151229986
#> log_rec_dev -0.245096006 0.5148710826
#> log_rec_dev -0.212008011 0.5193099393
#> log_rec_dev -0.198154382 0.5202926361
#> log_rec_dev -0.204956939 0.5180344854
#> log_rec_dev -0.229851806 0.5136491566
#> log_rec_dev -0.266246465 0.5078634363
#> log_rec_dev -0.296366985 0.5033009217
#> log_rec_dev -0.308546390 0.5013969915
#> log_rec_dev -0.313253299 0.5007385854
#> log_rec_dev -0.317285593 0.5002774307
#> log_rec_dev -0.312023245 0.5008781958
#> log_rec_dev -0.294176812 0.5032972075
#> log_rec_dev -0.270564738 0.5066421285
#> log_rec_dev -0.277463391 0.5060395835
#> log_rec_dev -0.269792298 0.5077518726
#> log_rec_dev -0.237509424 0.5129091877
#> log_rec_dev -0.206464938 0.5180295786
#> log_rec_dev -0.150575917 0.5270700571
#> log_rec_dev -0.067197023 0.5409461593
#> log_rec_dev -0.037345260 0.5472529435
#> log_rec_dev -0.033389386 0.5494843390
#> log_rec_dev -0.016755164 0.5531133701
#> log_rec_dev -0.015009131 0.5524478858
#> log_rec_dev  0.008961951 0.5546364874
#> log_rec_dev -0.037499332 0.5449332777
#> log_rec_dev -0.066625227 0.5377666594
#> log_rec_dev -0.127091247 0.5266060088
#> log_rec_dev -0.175924264 0.5177299322
#> log_rec_dev -0.195983618 0.5122340794
#> log_rec_dev -0.237248084 0.5042982837
#> log_rec_dev -0.267630154 0.4986282030
#> log_rec_dev -0.282631802 0.4955933053
#> log_rec_dev -0.311102517 0.4917555771
#> log_rec_dev -0.297848438 0.4928703264
#> log_rec_dev -0.241690588 0.4992099618
#> log_rec_dev -0.196253920 0.5050128722
#> log_rec_dev -0.177482641 0.5084187079
#> log_rec_dev -0.155910710 0.5113651636
#> log_rec_dev -0.128745159 0.5116720271
#> log_rec_dev -0.215279661 0.4980689722
#> log_rec_dev -0.313307306 0.4831124840
#> log_rec_dev -0.391961805 0.4723610506
#> log_rec_dev -0.421247706 0.4671835086
#> log_rec_dev -0.434532545 0.4637536451
#> log_rec_dev -0.488327616 0.4571295606
#> log_rec_dev -0.560649561 0.4491333922
#> log_rec_dev -0.630119659 0.4401942835
#> log_rec_dev -0.668442621 0.4327567101
#> log_rec_dev -0.696914695 0.4295885425
#> log_rec_dev -0.733071443 0.4248176310
#> log_rec_dev -0.795102762 0.4184384079
#> log_rec_dev -0.810623588 0.4140989913
#> log_rec_dev -0.826027785 0.4108981561
#> log_rec_dev -0.829316197 0.4105702743
#> log_rec_dev -0.728541821 0.4195898258
#> log_rec_dev -0.488687343 0.4381550818
#> log_rec_dev -0.146396756 0.4581390859
#> log_rec_dev  0.101642377 0.4453071163
#> log_rec_dev  0.441380554 0.3590746372
#> log_rec_dev -0.049113699 0.4938512740
#> log_rec_dev  1.435443072 0.1638096608
#> log_rec_dev -0.601908017 0.4651803521
#> log_rec_dev -0.693012692 0.4560706418
#> log_rec_dev  1.463038645 0.2130131176
#> log_rec_dev  0.839222876 0.6837709136
#> log_rec_dev  1.327537303 0.4130762052
#> log_rec_dev  0.128804587 0.5990360843
#> log_rec_dev  0.027014546 0.5540022525
#> log_rec_dev  0.165560756 0.6025115464
#> log_rec_dev  2.067427834 0.2393321570
#> log_rec_dev  0.590288513 0.9240650669
#> log_rec_dev  1.198961722 0.9607683341
#> log_rec_dev  0.488659366 0.8414113926
#> log_rec_dev  0.125367274 0.6371169766
#> log_rec_dev  0.000000000 0.6000000000
#> R0           0.978611174 0.0535766324
#> h            0.990000000 0.0000000000
#> q            0.008233652 0.0005347081
#> sigma        0.157646508 0.0101283422