Make a custom management procedure (MP)

Function operator that combines a function of class Assess and a function of class HCR to create a management procedure (MP). The resulting function can then be tested in closed-loop simulation via runMSE.

make_MP(.Assess, .HCR, diagnostic = c("min", "full", "none"), ...)

Arguments

.Assess
Assessment model, a function of class Assess.

.HCR
Harvest control rule, a function of class HCR.

diagnostic
A character string describing if any additional diagnostic information from the assessment models will be collected during the closed-loop simulation ("none" is the default). "min" (minimal) will collect information on convergence and "full" will also collect the model estimates of biomass and F generated by .Assess.

...
Additional arguments to be passed to .Assess and .HCR.

Value

A function of class MP.

See also

Examples

# A statistical catch-at-age model with a 40-10 control rule
SCA_40_10 <- make_MP(SCA, HCR40_10)

# An SCA that will produce convergence diagnostics
SCA_40_10 <- make_MP(SCA, HCR40_10, diagnostic = "min")

# MP with an SCA that uses a Ricker stock-recruit function.
SCA_Ricker <- make_MP(SCA, HCR_MSY, SR = "Ricker")
show(SCA_Ricker)

#> function (x, Data, reps = 1) 
#> {
#>     dependencies <- "Data@Cat, Data@Ind, Data@Mort, Data@L50, Data@L95, Data@CAA, Data@vbK, Data@vbLinf, Data@vbt0, Data@wla, Data@wlb, Data@MaxAge, Data@steep, Data@CV_Ind, Data@sigmaR, Data@CV_Cat, Data@Mort, Data@CV_Mort, Data@steep, Data@CV_steep, Data@vbLinf, Data@vbK, Data@vbt0, Data@wla, Data@wlb, Data@MaxAge, Data@L50, Data@L95"
#>     do_Assessment <- SCA(x = x, Data = Data, SR = "Ricker")
#>     Rec <- HCR_MSY(Assessment = do_Assessment, reps = reps)
#>     Rec@Misc <- Assess_diagnostic(x, Data, do_Assessment, include_assessment = FALSE)
#>     return(Rec)
#> }
#> <environment: 0x7f9e7d5bbaa8>
#> attr(,"class")
#> [1] "MP"