Plot Assessment object

Produces HTML file (via markdown) figures of parameter estimates and output from an Assessment object.

# S4 method for Assessment,missing
plot(
  x,
  filename = paste0("report_", x@Model),
  dir = tempdir(),
  ret_yr = 0L,
  open_file = TRUE,
  quiet = TRUE,
  render_args = list(),
  ...
)

# S4 method for Assessment,retro
plot(
  x,
  y,
  filename = paste0("report_", x@Model),
  dir = tempdir(),
  open_file = TRUE,
  quiet = TRUE,
  render_args = list(),
  ...
)

Arguments

x
An object of class Assessment.

filename
Character string for the name of the markdown and HTML files.

dir
The directory in which the markdown and HTML files will be saved.

ret_yr
If greater than zero, then a retrospective analysis will be performed and results will be reported. The integer here corresponds to the number of peels (the maximum number of terminal years for which the data are removed).

open_file
Logical, whether the HTML document is opened after it is rendered.

quiet
Logical, whether to silence the markdown rendering function.

render_args
Arguments to pass to render.

...
Other arguments.

y
An object of class retro.

Value

Returns invisibly the output from render.

See also

Examples

output <- DD_TMB(Data = Simulation_1)

# \donttest{
plot(output)

#> Writing markdown file: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpGZKwtj/report_DD_TMB.Rmd

#> Rendering markdown file...

#> Rendered file: /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpGZKwtj/report_DD_TMB.html

# }